GIT ERROR. This was supposed to be “AppStore version 1.0.2 - Added 64 bit support.” but turned out opencv2.framework just don’t work, so I deleted it and I will put it back later.

+++ Pocket Reader/ViewController.mm
@@ -598,7 +598,7 @@
     CGDataProviderRelease(provider);
     
     drawing.release();
-    const float blackMask[6] = { 0,0,0, 0,0,0 };
+    const CGFloat blackMask[6] = { 0,0,0, 0,0,0 };
     CGImageRef myColorMaskedImage = CGImageCreateWithMaskingColors(imageRef, blackMask);
     CGImageRelease(imageRef);
     
@@ -669,7 +669,7 @@
     CGColorSpaceRelease(colorSpace);
     CGDataProviderRelease(provider);
     image.release();
-    const float whiteMask[6] = { 255,255,255, 255,255,255 };
+    const CGFloat whiteMask[6] = { 255,255,255, 255,255,255 };
     CGImageRef myColorMaskedImage = CGImageCreateWithMaskingColors(imageRef, whiteMask);
     CGImageRelease(imageRef);
     dispatch_async(dispatch_get_main_queue(), ^{
